// reqy-web/lib/sdk-codegen/typescript-generator.ts
export interface GeneratedFile {
  path: string
  content: string
}